refactor setting context
